Move ExtractInfoFromCommit into package
