+++ internal/commitinfo/commitinfo_test.go
@@ -1,4 +1,4 @@
-package http
+package commitinfo
 
 import (
 	"testing"
@@ -11,13 +11,13 @@ func TestExtractInfoFromCommit(t *testing.T) {
 	tt := []struct {
 		name          string
 		commitMessage string
-		commitInfo    commitInfo
+		commitInfo    CommitInfo
 		err           error
 	}{
 		{
 			name:          "exact values",
 			commitMessage: "[test-service] artifact master-1234ds13g3-12s46g356g by Foo Bar\nArtifact-created-by: Foo Bar <test@lunar.app>",
-			commitInfo: commitInfo{
+			commitInfo: CommitInfo{
 				ArtifactID:  "master-1234ds13g3-12s46g356g",
 				AuthorEmail: "test@lunar.app",
 				AuthorName:  "Foo Bar",
@@ -28,7 +28,7 @@ func TestExtractInfoFromCommit(t *testing.T) {
 		{
 			name:          "email as author",
 			commitMessage: "[test-service] artifact master-1234ds13g3-12s46g356g by test@lunar.app\nArtifact-created-by: Foo Bar <test@lunar.app>",
-			commitInfo: commitInfo{
+			commitInfo: CommitInfo{
 				ArtifactID:  "master-1234ds13g3-12s46g356g",
 				AuthorEmail: "test@lunar.app",
 				AuthorName:  "Foo Bar",
@@ -39,19 +39,19 @@ func TestExtractInfoFromCommit(t *testing.T) {
 		{
 			name:          "not valid message",
 			commitMessage: "[product] build something",
-			commitInfo:    commitInfo{},
+			commitInfo:    CommitInfo{},
 			err:           errors.New("no match"),
 		},
 		{
 			name:          "release commit from product should not match",
 			commitMessage: "[dev/product] release test-s3-push-f4440b4ccb-1ba3085aa7 by eki@lunar.app\nArtifact-created-by: Emil Ingerslev <eki@lunar.app>\nArtifact-released-by: Bjørn Hald Sørensen <bso@lunar.app>",
-			commitInfo:    commitInfo{},
+			commitInfo:    CommitInfo{},
 			err:           errors.New("no match"),
 		},
 		{
 			name:          "artifact commit from product should match",
 			commitMessage: "[product] artifact test-s3-push-f4440b4ccb-1ba3085aa7 by eki@lunar.app\nArtifact-created-by: Emil Ingerslev <eki@lunar.app>",
-			commitInfo: commitInfo{
+			commitInfo: CommitInfo{
 				ArtifactID:  "test-s3-push-f4440b4ccb-1ba3085aa7",
 				AuthorEmail: "eki@lunar.app",
 				AuthorName:  "Emil Ingerslev",
@@ -62,7 +62,7 @@ func TestExtractInfoFromCommit(t *testing.T) {
 	}
 	for _, tc := range tt {
 		t.Run(tc.name, func(t *testing.T) {
-			info, err := extractInfoFromCommit()(tc.commitMessage)
+			info, err := ExtractInfoFromCommit()(tc.commitMessage)
 			if tc.err != nil {
 				assert.EqualError(t, errors.Cause(err), tc.err.Error(), "output error not as expected")
 			} else {
